test: refactor tests to reflect package structure
